Renaming of files and functions acc.to PR comments.
